# Make our test suite faster (again)