--- a/circle.yml
+++ b/circle.yml
@@ -18,7 +18,7 @@ dependencies:
 
 test:
   override:
-    - docker run fxa-auth-server:test npm test
+    - docker run fxa-auth-server:test npm run test-ci
 
 deployment:
   hub_latest: